Remove caffe import madness
Now we just say 'import caffe'. It is the callers responsibility to
ensure that this works. Since there is not default way to install
pycaffe, that means most likely that you have to 'sys.path.append' with
the caffe/python directory.

At least, this way you know what you are doing.