RandomLocation: Weaken GT containment criteria
Previously, only random locations were chosen that result in the output
ROI of the requested batch to be entirely inside the GT ROI. However,
this is not always possible (or desireable).

Now, locations are chosen such that the center of the output ROI is
within the GT roi (if a GT ROI is present).

This should similarly avoid too many rejects because of too little GT
masked in.

This commit also introduces Coordinate for easier coordinates/ROI
arithmetic.