Support user-defined volume types
Following change on master branch: 1379f40c08b51d12f827926089cc4df65bb3544e
This renames the previous 'VolumeType' to 'VolumeTypes' (which is a
collection of 'VolumeType' now, possibly created by the user).

'VolumeType' now also contains the 'interpolate' flag, so that sources
don't have to set it. The nodes that need it can query it directly from
'VolumeTypes'.