Fix bug in RandomLocation, int was expected
In gunpowder node RandomLocation, the begin and end of a ROI is used as an argument for the randint frunction, which requires an int as an argument. The datatype of begin and end of a ROI is not constrained to be int. Bug was fixed by converting the arguments to the randint function in place into int.