Make elastic augmentation of points more reliable
During elastic transformation, it might happen, that regions or voxels from the original image are removed in the transformed target image. This is a problem for 1-voxel large regions as currently used during point transformation. This commit iteratively enlarges the region of the points to increase the chances that the point also appears in the target image. The disappearance of points however should still be tolerated (for instance in cases where points are located at the border of the image).