Adapt elastic augment to points in worls units
Points of gunpowder were before assumed to be in voxel units. Now, it is assumed that points are in world units. In elastic augment, this requires some changes, since in order to augment the points, as a transient module, a volume is created. Points locations need to be adjusted accordingly in voxel units.