Bug fix in RasterizeNode for empty input case
Before this bug fix, RasterizeNode returned an incorrect map when no point at all was present in the input. For an empty input - no points at all in the current batch, one would expect a complete empty binary map. However (depending on the radius provided in the settings) a quarter blob was created always at the top right corner of the batch. This bug only occured when the parameter stay_inside_volumetype was set to None.