Add node for merging multiple upstream providers
So far, all requested volumes and points had to be provided by a single source. This node allows to merge volumes and/or points from different sources/ upstream providers. The context of adding this feature is to allow to have different sources for points and volumes. However, this node can for instance also be used for merging volumes from different sources (eg if the input volume is stored in a different file than the ground truth volume).