Fix bug when using mask in RasterizeNode
RasterizeNode requires a larger context area for rasterization. When using a mask, the requested roi for the mask is enlarged. When enlarging the area for mask, it has not been into account that the roi needs to be a multiple of the mask's voxel_size. The enlarged mask volume also needs to be recropped according to the upstream request.