Fixes bug when creating new array in rasterizenode
This commit fixes a bug, that happened when a sigma/radius was provided for rasterization that was not multiple of voxel size. During batch preparation, the points roi is enlarged by sigma to account for points lying outside the batch. During processing, the points roi is used to create the new array (and infer its shape). Inferring the roi for the new array has now been corrected. Before, it happened that the newly created array was  smaller than the points roi, leading to index error during rasterization. Now, the roi for the new array is snaped to voxel size.