Changes function to create ball structuring element
The original funtion used skimage.morphology to create a structuring element. This commit changes such that a distance transform is used (using enlarge_binary_map). The change is made in order to obtain the exact same mask as if enlarge_binary_map alone was used for rasterization (in order to maintain similar behaviour). For instance, a setting of radius=1 and a much larger voxel size should still produce a single voxel marked (versus returning an empty matrix).