Avoid is_alive() test on producer pool get()
get() should be callable from any process, not just the one that is
parent to the producer pool. Therefore, is_alive() should not be called
from within get(). Instead of testing whether the watchdog is still
alive, we now put an Exception into the result queue which will then be
available to the calling process.