Add option to store min/max of datasets in snapshots
This is useful when displaying a snapshot to infer an appropriate contrast range automatically.

I tried to store a dictionary {'valueRange': {'min': <min>, 'max': <max>}} but that does not seem to work with hdf5
If there is a way to actually do that in hdf5, please let me know.
Note: dumping as json string stores a string (and not a json/dictionary object)