Keep gradient info on tensors passed to loss
Since we have multiple mappings from function arguments to ArrayKeys it
is a bit awkward to pass tensors. The initial attempt simply added the
model outputs to the batch, and then gathered the loss inputs as was
done for model inputs. However in this transition the gradient
infomation was lost on model outputs.

Current approach gathers loss inputs from the batch and model seperately
and then merges the two.