Temporary solution to Graphs crop
When cropping out a graph, the naieve solution is to remove outside
nodes. In this case, using copy=False removes data from the original.
This breaks the BatchFilter crop -> process -> merge operation.

Temporary solution, copy the graph, subsample its nodes and edges,
and create a reference to the spec that can be changed.