Use returned batch from self.process
If available, use the returned value from process instead of the input
`node_batch`. Best practice is to return a batch with a nodes outputs,
but previous versions required inplace operation on `node_batch`. In
most cases this will still produce the correct results, but wherever the
dependencies include a crop out of a full dataset, inplace operations
will fail.