Fix ROI issue when using both transpose and mirror
SimpleAugment would request ROI's outside of the limits because the
total roi was not transposed when the actual roi was. This would create
incorrect offsets.

Swaping the order of transpose and mirror may also fix the issue,
but I did not know if they were in that order on purpose.