Replace BatchProviderTree with side effect free Pipeline
In contrast to BatchProviderTree, Pipeline objects do not modify
BatchProviders contained in them until the pipeline is build with
gp.build(pipeline). In addition, all changes made to BatchProviders
during the build are reverted when leaving the build context.

This allows BatchProviders and (partial) pipelines to be reused in
different build contexts. This is especially useful in interactive
environments like Jupyter notebooks, where BatchProviders are regularly
reused during prototyping.