Add JAX train/predict support
Notes:
- BufferPredict is factored out of the Tensorflow `Predict` and then
used as a parent class for JAX `Predict`. This should make using
BufferPredict easier for other implementations.
- `Train` was written based off PyTorch implementation. Since JAX is
rather low level and allowing many freedom for model implementation,
to make things easier for us, we make a simple `GenericJaxModel`
interface for models to follow in order to train or predict. A model
implementing this interface will need to contain not only the forward
model but also loss and update fn. Some examples can be found in
https://github.com/funkelab/funlib.learn.jax