Add JAX train/predict support (#169)
Some notes:
- Some examples of using JAX are in `tests/cases/torch_train.py`
- `Train` was written based off PyTorch implementation. Since JAX is
rather low level and allowing many freedom for model implementation,
to make things easier for us, we make a simple `GenericJaxModel`
interface for models to follow in order to train or predict. A model
implementing this interface will need to contain not only the forward
model but also loss and update fn. Some examples can be found in
https://github.com/funkelab/funlib.learn.jax