fix shift_augment test
test_pipeline2: Shift augment may increase the request roi, thus making
a request at the boundary of the provided roi may fail.
test_pipeline3: With a request size of 60 and a total roi of 100, this
leaves us 40 voxels of space for jitter. If `max(jitter) - min(jitter)`
is greater than 8 sigma this test fails. Dropping sigma to 4 means
we have a lower chance of failing this test