bugfix: DeformAugment point projection
Since we have a lookup of `target_location -> source_location`, when we
project points we find the source location that minimizes the distance
to the desired point, and then lookup the associated target location.
If a point is far ourside of our ROI, we will still find a "closest"
source location. To avoid this we were checking distances of neighboring
points to determine if a point landed out of bounds. However
this could lead to valid points being lost.
Now we simply grow the ROI by 1 voxel, project points, and then crop
away any nodes that were projected to the periphery.