fix (RandomLocation): ensure point voxel ROI is contained in original
To find possible shifts of all requested ROIs, the point request ROI has
to be converted into voxel space. The previous implementation did that
by a division, which rounds down. This can result in the voxel ROI not
being contained in the original request ROI, if the original did not
happen to line up with the voxel grid. This, in turn, can lead to a
situation where the randomly chosen point (that is supposed to be in the
point request ROI) is only contained in the voxel ROI, but not the
original ROI. And that triggers an assert further down.

Fixed now by snapping the point request ROI to the voxel grid by
shrinking before dividing by voxel size.