Generalize intensity augment to use slabs
Currently we support intensity shifts accross z slices, but we have a concept of slabs in other nodes providing much more freedom over which axes and step sizes get augmented independently.
This change depricates the z_section_wise argument in favor of slabs, but keeps support for both for now.